backend/pyproject.toml
[project]
name = "book-embeddings"
version = "0.1.0"
description = "Backend system for crawling Docusaurus URLs, generating embeddings using Cohere models, and storing them in Qdrant Cloud"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=4.0.0",
    "qdrant-client>=1.9.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"